20220805  继承规则遗留问题数据处理逻辑调整及脚本

+++ src/main/java/com/qtgl/iga/service/impl/NodeRulesServiceImpl.java
@@ -74,10 +74,12 @@ public class NodeRulesServiceImpl implements NodeRulesService {
         //删除rules
         Integer flag = nodeRulesDao.deleteNodeRulesById((String) arguments.get("id"));
         //查询node
-        List<NodeRulesVo> nodeRulesByNodeId = nodeRulesDao.findNodeRulesByNodeId(nodeRules.getNodeId(), null);
-        //node下没有对应的nodeRules则将node删除
-        if (CollectionUtils.isEmpty(nodeRulesByNodeId)) {
-            nodeDao.deleteNodeById(nodeRules.getNodeId(), domain);
+        if (null != nodeRules) {
+            List<NodeRulesVo> nodeRulesByNodeId = nodeRulesDao.findNodeRulesByNodeId(nodeRules.getNodeId(), 1);
+            //node下没有对应的nodeRules则将node删除
+            if (CollectionUtils.isEmpty(nodeRulesByNodeId)) {
+                nodeDao.deleteNodeById(nodeRules.getNodeId(), domain);
+            }
         }
         if (flag > 0) {
             return new NodeRulesVo();
